partB_func_coach/requirements.txt
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.34.0
azure-functions==1.23.0
azure-identity==1.23.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
packaging==25.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
